Create pubkeyToAddress in @cosmjs/amino
